pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "jcloud"
version = "0.1.0"
description = "JumpCloud management CLI"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "click>=8.1",
    "tabulate>=0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
]

[project.scripts]
jcloud = "jcloud.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
